[rename] #106 clip detail -> link 네이밍 변경
