+++ feature/clip/src/main/java/org/sopt/clip/cliplink/ClipLinkViewHolder.kt
@@ -1,12 +1,12 @@
-package org.sopt.clip.clipdetail
+package org.sopt.clip.cliplink
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.clip.LinkDTO
-import org.sopt.clip.databinding.ItemClipDetailLinkBinding
+import org.sopt.clip.databinding.ItemClipLinkBinding
 import org.sopt.ui.view.onThrottleClick
 
 class ClipLinkViewHolder(
-  private val binding: ItemClipDetailLinkBinding,
+  private val binding: ItemClipLinkBinding,
   private val onClickItemLink: (Long, String) -> Unit,
 ) : RecyclerView.ViewHolder(binding.root) {
   fun onBind(linkData: LinkDTO, onClick: (LinkDTO) -> Unit) {
