+++ feature/clip/src/main/res/layout/item_clip_link.xml
@@ -35,6 +35,7 @@
     app:layout_constraintTop_toTopOf="parent" />
 
   <ImageView
+    android:id="@+id/iv_link_delete"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:layout_margin="-12dp"
